Use processor.cpp API in disasm.cpp
Remove ARM specific workaround and do not assume the CPU name contains all
information about the code.
This should now allow disassemble of code that's not required by the base ISA
or even not supported on the current CPU.
This is especially important on ARM, which does not disassemble many armv7 instructions previously.