inference: create a separate type for doing optimizations
Most fields of InferenceState aren't valid during optimization,
so the goal is to reflect that in the structure of the types

Similarly, doing optimization operations during inference would be
invalid, so this helps distinguish those cases as well.

Finally, this makes is possible for non-InferenceState-initiated IR passes
(e.g. external to typeinf) to make uses of these passes.